Bug/Fix update -
Fixing an issue with the condaEnv.bat script not finding the environment that was imported from the file. Need a try/catch for the install to speed up activation.

Commented out a bugged argument in the shRemote.py file. Tagged in comments.

Adding requireents.txt so this can be installed with only pip, and no requirement on anaconda (even though anaconda is prefered)